+++ ArtemisExamCheckKit/Sources/ArtemisExamCheckKit/Views/StudentDetailView/StudentDetailView.swift
@@ -134,65 +134,7 @@ struct StudentDetailView: View {
                 }
                 .padding(.vertical, 16)
 
-                Group {
-                    if showSigningImage {
-                        HStack(alignment: .bottom) {
-                            ArtemisAsyncImage(
-                                imageURL: student.signingImageURL,
-                                onFailure: { signingImageLoadingStatus = .failure(error: $0) },
-                                onProgress: { _, _ in signingImageLoadingStatus = .loading },
-                                onSuccess: { _ in signingImageLoadingStatus = .success }
-                            ) {
-                                HStack {
-                                    Spacer()
-                                    VStack {
-                                        Image(systemName: "signature")
-                                            .resizable()
-                                            .aspectRatio(contentMode: .fit)
-                                            .foregroundColor(.red)
-                                        Text("The signature could not be loaded :(")
-                                            .font(.caption)
-                                            .foregroundColor(.red)
-                                    }.frame(height: 200)
-                                    Spacer()
-                                }
-                            }
-                            .scaledToFit()
-                            .frame(height: 200)
-                            switch signingImageLoadingStatus {
-                            case .notStarted, .loading:
-                                EmptyView()
-                            case .success, .failure:
-                                PencilSideButtons(isScrollingEnabled: $isScrollingEnabled,
-                                                  student: $student,
-                                                  showSigningImage: $showSigningImage,
-                                                  canvasView: $canvasView)
-                            }
-                        }
-                    } else {
-                        HStack(alignment: .bottom) {
-                            CanvasView(canvasView: $canvasView)
-                                .frame(minHeight: 200)
-                                .border(Color(UIColor.label))
-                            VStack(spacing: 32) {
-                                Button {
-                                    isScrollingEnabled.toggle()
-                                } label: {
-                                    Image(systemName: "hand.draw.fill")
-                                        .imageScale(.large)
-                                        .foregroundColor(isScrollingEnabled ? Color.gray : Color.blue)
-                                }
-                                Button {
-                                    canvasView.drawing = PKDrawing()
-                                } label: {
-                                    Image(systemName: "trash.fill")
-                                        .imageScale(.large)
-                                        .foregroundColor(.red)
-                                }
-                            }
-                        }
-                    }
-                }
+                signingImageOrCanvas
 
                 Button("Save") {
                     saveStudent()
@@ -253,105 +195,62 @@ struct StudentDetailView: View {
         }
     }
 
-    private func saveStudent(force: Bool = false, isNavigationBarButton: Bool = false) {
-        if !force && (!didCheckName || !didCheckLogin || !didCheckImage || !didCheckRegistrationNumber || (canvasView.drawing.bounds.isEmpty && student.signingImageURL == nil)) {
-            if isNavigationBarButton {
-                showDidNotCompleteDialogNavigationBar = true
-            } else {
-                showDidNotCompleteDialog = true
+    @ViewBuilder private var signingImageOrCanvas: some View {
+        if showSigningImage {
+            HStack(alignment: .bottom) {
+                ArtemisAsyncImage(
+                    imageURL: student.signingImageURL,
+                    onFailure: { signingImageLoadingStatus = .failure(error: $0) },
+                    onProgress: { _, _ in signingImageLoadingStatus = .loading },
+                    onSuccess: { _ in signingImageLoadingStatus = .success }
+                ) {
+                    HStack {
+                        Spacer()
+                        VStack {
+                            Image(systemName: "signature")
+                                .resizable()
+                                .aspectRatio(contentMode: .fit)
+                                .foregroundColor(.red)
+                            Text("The signature could not be loaded :(")
+                                .font(.caption)
+                                .foregroundColor(.red)
+                        }.frame(height: 200)
+                        Spacer()
+                    }
+                }
+                .scaledToFit()
+                .frame(height: 200)
+                switch signingImageLoadingStatus {
+                case .notStarted, .loading:
+                    EmptyView()
+                case .success, .failure:
+                    PencilSideButtons(isScrollingEnabled: $isScrollingEnabled,
+                                      student: $student,
+                                      showSigningImage: $showSigningImage,
+                                      canvasView: $canvasView)
+                }
             }
-            return
-        }
-
-        var imageData: Data?
-        if !canvasView.drawing.bounds.isEmpty {
-            let signingImage = canvasView.drawing.image(from: canvasView.bounds, scale: UIScreen.main.scale)
-            imageData = signingImage.pngData()
-        }
-
-        let newStudent = student.copy(checkedImage: didCheckImage,
-                                      checkedName: didCheckName,
-                                      checkedLogin: didCheckLogin,
-                                      checkedRegistrationNumber: didCheckRegistrationNumber,
-                                      actualRoom: actualOtherRoom.isEmpty ? (actualRoom.isEmpty ? nil : actualRoom) : actualOtherRoom,
-                                      actualSeat: actualSeat.isEmpty ? nil : actualSeat,
-                                      signing: imageData)
-
-        // format for name <examId>-<examUserId>-<examUserName>-<registrationNumber>.png
-        let imageName = "\(examId)-\(student.id)-\(student.user.name)-\(student.user.visibleRegistrationNumber ?? "missing").png"
-        saveImageToDocuments(imageData: imageData, imageName: imageName)
-
-        Task {
-            isSaving = true
-            let result = await StudentServiceFactory.shared.saveStudent(student: newStudent, examId: examId, courseId: courseId)
-            switch result {
-            case .loading:
-                isSaving = true
-            case .failure(let error):
-                isSaving = false
-                self.error = error
-            case .done(let newStudent):
-                isSaving = false
-                student = newStudent
-                updateDetailViewStates()
-                await successfullySavedCompletion(newStudent)
-            }
-        }
-    }
-
-    private func updateDetailViewStates() {
-        didCheckImage = student.didCheckImage ?? false
-        didCheckName = student.didCheckName ?? false
-        didCheckLogin = student.didCheckLogin ?? false
-        didCheckRegistrationNumber = student.didCheckRegistrationNumber ?? false
-        showSigningImage = student.signingImageURL != nil
-        actualRoom = student.actualRoom ?? ""
-        actualSeat = student.actualSeat ?? ""
-    }
-}
-
-// MARK: - File Manager
-
-private extension StudentDetailView {
-
-    func saveImageToDocuments(imageData: Data?, imageName: String) {
-        guard let data = imageData,
-              let documentsDirectory = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first else { return }
-
-        let fileURL = documentsDirectory
-            .appendingPathComponent("ExamAttendaceChecker")
-            .appendingPathComponent(imageName)
-
-        createDirectoryIfNecessary()
-
-        // Checks if file exists, removes it if so.
-        if FileManager.default.fileExists(atPath: fileURL.path) {
-            do {
-                try FileManager.default.removeItem(atPath: fileURL.path)
-                print("Removed old image")
-            } catch {
-                print("couldn't remove file at path", error)
-            }
-        }
-
-        do {
-            try data.write(to: fileURL)
-        } catch {
-            print("error saving file with error", error)
-        }
-    }
-
-    func createDirectoryIfNecessary() {
-        guard let documentsDirectory = FileManager.default.urls(for: .documentDirectory, in: .userDomainMask).first else { return }
-
-        let folderURL = documentsDirectory
-            .appendingPathComponent("ExamAttendaceChecker")
-
-        if !FileManager.default.fileExists(atPath: folderURL.path) {
-            do {
-                try FileManager.default.createDirectory(atPath: folderURL.path, withIntermediateDirectories: true, attributes: nil)
-            } catch {
-                print(error.localizedDescription)
+        } else {
+            HStack(alignment: .bottom) {
+                CanvasView(canvasView: $canvasView)
+                    .frame(minHeight: 200)
+                    .border(Color(UIColor.label))
+                VStack(spacing: 32) {
+                    Button {
+                        isScrollingEnabled.toggle()
+                    } label: {
+                        Image(systemName: "hand.draw.fill")
+                            .imageScale(.large)
+                            .foregroundColor(isScrollingEnabled ? Color.gray : Color.blue)
+                    }
+                    Button {
+                        canvasView.drawing = PKDrawing()
+                    } label: {
+                        Image(systemName: "trash.fill")
+                            .imageScale(.large)
+                            .foregroundColor(.red)
+                    }
+                }
             }
         }
     }
